comment(쓰기, 수정, 삭제, 평점 반영)

+++ src/main/java/com/project/board/to/CommentTO.java
@@ -18,7 +18,7 @@ public class CommentTO {
 	private int cmtGrps;
 	private int cmtGrpl;
 	private String cmtContent;
-	private int cmtGrade;
+	private double cmtGrade;
 	private Date cmtRegDate;
 	
 }
